Revert "[WIP] PDBx/mmcif support"
