+++ PBlib.py
@@ -81,25 +81,62 @@ FASTA_WIDTH = 60
 #===============================================================================
 # Functions
 #===============================================================================
+def get_dihedral(atomA, atomB, atomC, atomD):
+    """
+    Compute dihedral angle between 4 atoms (A, B, C, D)
+    each atom is represented as a list of three coordinates [x, y, z]
+    output is in degree in the range -180 / +180
+    """
+    
+    # convert lists to Numpy objects
+    A = numpy.array(atomA)
+    B = numpy.array(atomB)
+    C = numpy.array(atomC)
+    D = numpy.array(atomD)
+ 
+    # vectors
+    AB = B - A 
+    BC = C - B 
+    CD = D - C 
 
+    # normal vectors
+    n1 = numpy.cross(AB, BC)
+    n2 = numpy.cross(BC, CD)
+
+    # normalize normal vectors
+    n1 /= numpy.linalg.norm(n1)
+    n2 /= numpy.linalg.norm(n2)
+    
+    # angle between normals
+    cosine = numpy.sum(n1*n2) / (numpy.linalg.norm(n1) * numpy.linalg.norm(n2))
+    try :
+        torsion = math.acos(cosine)
+    except:
+        cosine = int(cosine) #+0.0001
+        torsion = math.acos(cosine)
+
+    # convert radion to degree
+    torsion = torsion * 180.0 / math.pi 
+
+    # find if the torsion is clockwise or counterclockwise
+    #if numpy.sum(n1 * CD) < 0.0:
+    if numpy.dot(n1, CD) < 0.0:
+        torsion = 360 - torsion
+    if torsion == 360.0:
+        torsion = 0.0
+    
+    # get range -180 / +180
+    if torsion > 180.0:
+        torsion = torsion - 360
+    if torsion < -180.0:
+        torsion = torsion + 360
+   
+    return torsion
 
 #-------------------------------------------------------------------------------    
 def read_fasta(name):
     """
-    Read fasta file and output sequences in a list.
-    
-    Parameters
-    ----------
-    name : str
-        Name of file containing sequences in fasta format.
-    
-    Returns
-    -------
-    header_lst : list
-        List of headers (str)
-    sequence_lst : list
-        List of sequences (str)
-    
+    Read fasta file and output sequences in a list
     """
     assert os.path.exists(name), name + ' does not exist'
     sequence_lst = []
@@ -141,19 +178,10 @@ def read_fasta(name):
 #-------------------------------------------------------------------------------
 def load_substitution_matrix(name):
     """
-    Load PB substitution matrix.
-    
-    Parameters
-    ----------
-    name : str
-        Name of the file containing the PBs susbtitution matrix.
-    
-    Returns
-    -------
-    mat : numpy array
-        Array of floats.
+    Load PB substitution matrix
     """
     try:
+        # mat = numpy.loadtxt(name, dtype=int, skiprows=2)
         mat = numpy.loadtxt(name, dtype=float, skiprows=2)
     except:
         sys.exit("ERROR: cannot read %s" % name)
@@ -171,12 +199,7 @@ def load_substitution_matrix(name):
 #-------------------------------------------------------------------------------
 def clean_file(name):
     """
-    Clean existing file.
-    
-    Parameters
-    ----------
-    name : str
-        Name of file to remove.
+    Clean existing file
     """
     if os.path.exists(name):
         os.remove(name)
@@ -184,17 +207,8 @@ def clean_file(name):
 #-------------------------------------------------------------------------------
 def write_fasta(name, seq, comment):
     """
-    Format seq and comment to fasta format and write file.
-    
-    Parameters
-    ----------
-    name : str
-        Name of file to write.
-    seq : str
-        Sequence to format.
-    comment : str
-        Comment to make header of sequence.
-    
+    Format seq and comment to fasta format
+    and write file
     """
     fasta_content  = ">"+comment+"\n"
     fasta_content += "\n".join( [seq[i:i+FASTA_WIDTH] for i in range(0, len(seq), FASTA_WIDTH)] )
