+++ test_data/1BTA.PB.fasta
