--- test_data/1BTA.cif.gz.PB.fasta
@@ -1,3 +1,0 @@
->test_data/1BTA.cif.gz | model 1 | chain A
-ZZdddfklonbfklmmmmmmmmnopafklnoiaklmmmmmnoopacddddddehkllmmm
-mngoilmmmmmmmmmmmmnopacdcddZZ
