--- test_data/1BTA.cif.gz.PB.phipsi
@@ -1,89 +1,0 @@
-test_data/1BTA.cif.gz | model 1 | chain A      1     None  -171.66 
-test_data/1BTA.cif.gz | model 1 | chain A      2  -133.80   153.74 
-test_data/1BTA.cif.gz | model 1 | chain A      3  -134.66   157.30 
-test_data/1BTA.cif.gz | model 1 | chain A      4  -144.49   118.60 
-test_data/1BTA.cif.gz | model 1 | chain A      5  -100.13    92.99 
-test_data/1BTA.cif.gz | model 1 | chain A      6   -83.49   104.24 
-test_data/1BTA.cif.gz | model 1 | chain A      7   -64.77   -43.25 
-test_data/1BTA.cif.gz | model 1 | chain A      8   -44.48   -25.89 
-test_data/1BTA.cif.gz | model 1 | chain A      9   -94.91   -47.18 
-test_data/1BTA.cif.gz | model 1 | chain A     10   -41.31   133.74 
-test_data/1BTA.cif.gz | model 1 | chain A     11  -119.15   -11.83 
-test_data/1BTA.cif.gz | model 1 | chain A     12  -174.21   175.87 
-test_data/1BTA.cif.gz | model 1 | chain A     13   -56.61   -45.75 
-test_data/1BTA.cif.gz | model 1 | chain A     14   -50.78   -45.37 
-test_data/1BTA.cif.gz | model 1 | chain A     15   -57.94   -43.33 
-test_data/1BTA.cif.gz | model 1 | chain A     16   -55.21   -56.48 
-test_data/1BTA.cif.gz | model 1 | chain A     17   -64.52   -18.58 
-test_data/1BTA.cif.gz | model 1 | chain A     18   -70.24   -55.15 
-test_data/1BTA.cif.gz | model 1 | chain A     19   -65.21   -41.28 
-test_data/1BTA.cif.gz | model 1 | chain A     20   -58.99   -35.79 
-test_data/1BTA.cif.gz | model 1 | chain A     21   -66.87   -42.15 
-test_data/1BTA.cif.gz | model 1 | chain A     22   -67.34   -57.40 
-test_data/1BTA.cif.gz | model 1 | chain A     23   -52.30   -66.09 
-test_data/1BTA.cif.gz | model 1 | chain A     24   -61.19   -14.81 
-test_data/1BTA.cif.gz | model 1 | chain A     25    54.95    47.60 
-test_data/1BTA.cif.gz | model 1 | chain A     26   -69.52   161.11 
-test_data/1BTA.cif.gz | model 1 | chain A     27   -57.36  -179.66 
-test_data/1BTA.cif.gz | model 1 | chain A     28   -79.91   -18.49 
-test_data/1BTA.cif.gz | model 1 | chain A     29   -93.52    13.80 
-test_data/1BTA.cif.gz | model 1 | chain A     30   -38.41   105.85 
-test_data/1BTA.cif.gz | model 1 | chain A     31   -64.02    -5.51 
-test_data/1BTA.cif.gz | model 1 | chain A     32    50.07     3.60 
-test_data/1BTA.cif.gz | model 1 | chain A     33   -84.84  -176.05 
-test_data/1BTA.cif.gz | model 1 | chain A     34   -76.66   -36.89 
-test_data/1BTA.cif.gz | model 1 | chain A     35   -66.21   -36.19 
-test_data/1BTA.cif.gz | model 1 | chain A     36   -80.77   -55.89 
-test_data/1BTA.cif.gz | model 1 | chain A     37   -45.10   -50.82 
-test_data/1BTA.cif.gz | model 1 | chain A     38   -58.51   -56.43 
-test_data/1BTA.cif.gz | model 1 | chain A     39   -44.01   -26.06 
-test_data/1BTA.cif.gz | model 1 | chain A     40   -79.68   -51.38 
-test_data/1BTA.cif.gz | model 1 | chain A     41   -58.81   -49.46 
-test_data/1BTA.cif.gz | model 1 | chain A     42   -75.73     3.92 
-test_data/1BTA.cif.gz | model 1 | chain A     43  -177.15    60.46 
-test_data/1BTA.cif.gz | model 1 | chain A     44   177.13   -66.63 
-test_data/1BTA.cif.gz | model 1 | chain A     45   -58.44   149.60 
-test_data/1BTA.cif.gz | model 1 | chain A     46  -102.66   132.43 
-test_data/1BTA.cif.gz | model 1 | chain A     47  -114.52   169.33 
-test_data/1BTA.cif.gz | model 1 | chain A     48   -61.39   136.70 
-test_data/1BTA.cif.gz | model 1 | chain A     49  -113.18   156.54 
-test_data/1BTA.cif.gz | model 1 | chain A     50  -117.26   138.51 
-test_data/1BTA.cif.gz | model 1 | chain A     51  -120.03    81.76 
-test_data/1BTA.cif.gz | model 1 | chain A     52   -77.61    83.18 
-test_data/1BTA.cif.gz | model 1 | chain A     53   -79.66   111.40 
-test_data/1BTA.cif.gz | model 1 | chain A     54  -100.37   150.03 
-test_data/1BTA.cif.gz | model 1 | chain A     55    49.87    68.74 
-test_data/1BTA.cif.gz | model 1 | chain A     56   -73.88   -66.74 
-test_data/1BTA.cif.gz | model 1 | chain A     57   -56.21   -35.21 
-test_data/1BTA.cif.gz | model 1 | chain A     58   -66.38   -32.22 
-test_data/1BTA.cif.gz | model 1 | chain A     59   -94.68    17.69 
-test_data/1BTA.cif.gz | model 1 | chain A     60  -111.49   -38.10 
-test_data/1BTA.cif.gz | model 1 | chain A     61   -70.65   -62.86 
-test_data/1BTA.cif.gz | model 1 | chain A     62   -33.51   -32.03 
-test_data/1BTA.cif.gz | model 1 | chain A     63  -128.57    62.58 
-test_data/1BTA.cif.gz | model 1 | chain A     64   -12.37   106.99 
-test_data/1BTA.cif.gz | model 1 | chain A     65    73.69    32.13 
-test_data/1BTA.cif.gz | model 1 | chain A     66   -89.05   -69.17 
-test_data/1BTA.cif.gz | model 1 | chain A     67   -77.83   -21.56 
-test_data/1BTA.cif.gz | model 1 | chain A     68   -71.32   -21.86 
-test_data/1BTA.cif.gz | model 1 | chain A     69   -81.41   -55.29 
-test_data/1BTA.cif.gz | model 1 | chain A     70   -52.05   -43.23 
-test_data/1BTA.cif.gz | model 1 | chain A     71   -59.22   -45.28 
-test_data/1BTA.cif.gz | model 1 | chain A     72   -52.67   -38.13 
-test_data/1BTA.cif.gz | model 1 | chain A     73   -75.01   -30.84 
-test_data/1BTA.cif.gz | model 1 | chain A     74   -74.70   -35.04 
-test_data/1BTA.cif.gz | model 1 | chain A     75   -80.23   -37.27 
-test_data/1BTA.cif.gz | model 1 | chain A     76   -63.33   -46.74 
-test_data/1BTA.cif.gz | model 1 | chain A     77   -62.58   -38.84 
-test_data/1BTA.cif.gz | model 1 | chain A     78   -58.44   -30.93 
-test_data/1BTA.cif.gz | model 1 | chain A     79   -77.26   -28.81 
-test_data/1BTA.cif.gz | model 1 | chain A     80   -65.77    -6.59 
-test_data/1BTA.cif.gz | model 1 | chain A     81   113.27   -14.07 
-test_data/1BTA.cif.gz | model 1 | chain A     82   -63.86   160.46 
-test_data/1BTA.cif.gz | model 1 | chain A     83  -109.29    65.33 
-test_data/1BTA.cif.gz | model 1 | chain A     84   -94.30    87.93 
-test_data/1BTA.cif.gz | model 1 | chain A     85   -52.92    98.90 
-test_data/1BTA.cif.gz | model 1 | chain A     86   -73.45   114.65 
-test_data/1BTA.cif.gz | model 1 | chain A     87  -114.16   101.25 
-test_data/1BTA.cif.gz | model 1 | chain A     88   -96.79   106.74 
-test_data/1BTA.cif.gz | model 1 | chain A     89  -109.03     None
