--- test_data/1BTA.pdb.PB.flat
@@ -1,1 +1,0 @@
-ZZdddfklonbfklmmmmmmmmnopafklnoiaklmmmmmnoopacddddddehkllmmmmngoilmmmmmmmmmmmmnopacdcddZZ
