--- test_data/3ICH.cif.gz.PB.fasta
@@ -1,4 +1,0 @@
->test_data/3ICH.cif.gz | model 1 | chain A
-ZZccdfbdcdddddehjbdebjcdddddfklmmmlmmmmmmmmnopnopajeopacfbdc
-ehibacehiamnonopgocdfkbjbdcdfblmbccfbghiacdddebehiafkbccddfb
-dcfklgokaccfbdcfbhklmmmmmmmpccdfkopafbacddfbgcddddfbacddddZZ
