--- test_data/all.cif.gz.PB.fasta
@@ -1,53 +1,0 @@
->test_data/1BTA.cif.gz | model 1 | chain A
-ZZdddfklonbfklmmmmmmmmnopafklnoiaklmmmmmnoopacddddddehkllmmm
-mngoilmmmmmmmmmmmmnopacdcddZZ
->test_data/1AY7.cif.gz | chain A
-ZZbjadfklmcfklmmmmmmmmnnpaafbfkgopacehlnomaccddehjaccdddddeh
-klpnbjadcdddfbehiacddfegolaccdddfkZZ
->test_data/1AY7.cif.gz | model 1 | chain B
-ZZcddfklpcbfklmmmmmmmmnopafklgoiaklmmmmmmmmpacddddddehkllmmm
-mnnommmmmmmmmmmmmmnopacddddZZ
->test_data/2LFU.cif.gz | model 1 | chain A
-ZZbghiacfkbccdddddehiadddddddddddfklggcdddddddddddddehifbdcd
-dddddddddfklopadddddfhpamlnopcddddddehjadddddehjacbddddddddf
-klmaccddddddfbgniaghiapaddddddfklnoambZZ
->test_data/2LFU.cif.gz | model 2 | chain A
-ZZpcfblcffbccdddddeehjacdddddddddfklggcddddddddddddddfblghia
-dddddddddfklopadddddehpmmmnopcddddddeehiacdddfblopadcddddddf
-klpaccdddddfklmlmgcdehiaddddddfklmmgopZZ
->test_data/2LFU.cif.gz | model 3 | chain A
-ZZmgghiafbbccdddddehjbdcdddddddddfklggcddddddddddddddfbfghpa
-cddddddddfklopadddddehiaklmmmgcdddddeehiaddddfkbgciacdddddef
-klpaccddddddfkgojbdfehpaddddddfkbccfbgZZ
->test_data/2LFU.cif.gz | model 4 | chain A
-ZZcghiacfkbacdddddfbhpacdddddddddfklmcfdddddddddddddehiacddd
-dddddddddfknopadddddfkpamlnopaddddddehjaccdddfklnopacddddddf
-klmpccdddddddehiabghehiaddddddfklpccfkZZ
->test_data/2LFU.cif.gz | model 5 | chain A
-ZZpaehiehkaccdddddehjbccdddddddddfklggcddddddddddddddfbhpadd
-dddddddddfklopadddddehiamlmmpccdddddeehiadddddfbacddcddddddf
-klmaccddddddfbgghiafehiadddddddfklpacfZZ
->test_data/2LFU.cif.gz | model 6 | chain A
-ZZmghbacfkbccdddddeehpacdddddddddfklggcdddddddddddddehiacadd
-dddddddddfklopadddddehiaklnopcddddddeehiadddehjlnopacddddddf
-klmaccddddehiaehbgcdehiadddddddfehjlpcZZ
->test_data/2LFU.cif.gz | model 7 | chain A
-ZZcchbacfkbccdddddfehpacdddddddddfklggcdddddddddddddddehjapa
-dddddddddfknopadddddfklmmmnopcddddddehjiddddddfknopacddddddf
-klpaccdddddfklmaacdfehpadddddehjblckknZZ
->test_data/2LFU.cif.gz | model 8 | chain A
-ZZcehjdeehiacdjdddedjbdcdddddddddfklggcdddddddddddddddbfblba
-cddddddddfklopacddddehiamlnopaddddddehjacddddfehpaaccdddddef
-klpaccdddddfklmbfbehehiaddddddffkgoiehZZ
->test_data/2LFU.cif.gz | model 9 | chain A
-ZZpccdjdfkbccdddddehhpacdddddddddfklggcdddddddddddddehiacbdc
-dddddddddfklopadddddehiammnopcddddddeejiadddehjlgobacddddddf
-klmpccddddehiacbcbdfehpadddddehjklmklmZZ
->test_data/2LFU.cif.gz | model 10 | chain A
-ZZccfklcfkbccdddddehjbdcdddddddddfklggcdddddddddddddehiapacc
-dddddddddfklopadddddehjamlnopaddddddehjddcdddfbfghpacddddddf
-klpaccddddddfbcfbacfehpadddddddekpghiaZZ
->test_data/3ICH.cif.gz | model 1 | chain A
-ZZccdfbdcdddddehjbdebjcdddddfklmmmlmmmmmmmmnopnopajeopacfbdc
-ehibacehiamnonopgocdfkbjbdcdfblmbccfbghiacdddebehiafkbccddfb
-dcfklgokaccfbdcfbhklmmmmmmmpccdfkopafbacddfbgcddddfbacddddZZ
